Endpoints/simpleApp/requirements.txt
Flask
pandas
matplotlib
seaborn
azure-storage-blob

./package/model_warehouse-2.0.0-py3-none-any.whl